Added variable-length delay line by sine wave
Read cursor follows writeCursor with outCurrDelay positions of delay
Added alignReadCursor to fit readCursor in range 0-delayMaxSize
Added calcOldestSample for interpolation
Added rate parameter in VST plugin

+++ src/ModOperator.cpp
@@ -20,7 +20,7 @@ Oscillator *ModOperator::getOscillator(){
     return oscillator;
 }
 //setters
-void ModOperator::setMinAmount(int minAmount){
+void ModOperator::setMinAmount(float minAmount){
     this->minAmount = minAmount;
 }
